Refactor requirement generators to use a common rule helper.
