+++ src/python/pants/backend/python/macros/python_requirements.py
@@ -3,41 +3,26 @@
 
 from __future__ import annotations
 
-import itertools
-import os
-from typing import Iterable
-
-from packaging.utils import canonicalize_name as canonicalize_project_name
+from typing import Iterator
 
 from pants.backend.python.macros.common_fields import (
     ModuleMappingField,
     RequirementsOverrideField,
     TypeStubsModuleMappingField,
 )
+from pants.backend.python.macros.common_requirements_rule import _generate_requirements
 from pants.backend.python.pip_requirement import PipRequirement
 from pants.backend.python.subsystems.setup import PythonSetup
 from pants.backend.python.target_types import (
-    PythonRequirementModulesField,
     PythonRequirementResolveField,
-    PythonRequirementsField,
     PythonRequirementTarget,
-    PythonRequirementTypeStubModulesField,
     parse_requirements_file,
 )
-from pants.core.target_types import (
-    TargetGeneratorSourcesHelperSourcesField,
-    TargetGeneratorSourcesHelperTarget,
-)
-from pants.engine.addresses import Address
-from pants.engine.fs import DigestContents, GlobMatchErrorBehavior, PathGlobs
-from pants.engine.internals.target_adaptor import TargetAdaptor, TargetAdaptorRequest
-from pants.engine.rules import Get, collect_rules, rule
+from pants.engine.rules import collect_rules, rule
 from pants.engine.target import (
     COMMON_TARGET_FIELDS,
-    Dependencies,
     GeneratedTargets,
     GenerateTargetsRequest,
-    InvalidFieldException,
     SingleSourceField,
     TargetGenerator,
 )
@@ -89,103 +74,17 @@ async def generate_from_python_requirement(
     union_membership: UnionMembership,
     python_setup: PythonSetup,
 ) -> GeneratedTargets:
-    generator = request.generator
-    requirements_rel_path = generator[PythonRequirementsSourceField].value
-    requirements_full_path = generator[PythonRequirementsSourceField].file_path
-    overrides = {
-        canonicalize_project_name(k): v
-        for k, v in request.require_unparametrized_overrides().items()
-    }
-
-    file_tgt = TargetGeneratorSourcesHelperTarget(
-        {TargetGeneratorSourcesHelperSourcesField.alias: requirements_rel_path},
-        Address(
-            request.template_address.spec_path,
-            target_name=request.template_address.target_name,
-            relative_file_path=requirements_rel_path,
-        ),
+    result = await _generate_requirements(
+        request,
         union_membership,
+        python_setup,
+        parse_requirements_callback=parse_requirements_callback,
     )
+    return GeneratedTargets(request.generator, result)
 
-    req_deps = [file_tgt.address.spec]
 
-    resolve = request.template.get(
-        PythonRequirementResolveField.alias, python_setup.default_resolve
-    )
-    lockfile = python_setup.resolves.get(resolve) if python_setup.enable_resolves else None
-    if lockfile:
-        target_adaptor = await Get(
-            TargetAdaptor,
-            TargetAdaptorRequest(
-                description_of_origin=f"python_requirements lockfile dep for the {resolve} resolve",
-                address=Address(
-                    os.path.dirname(lockfile),
-                    target_name=resolve,
-                ),
-            ),
-        )
-        if target_adaptor.type_alias == "_lockfiles":
-            req_deps.append(f"{lockfile}:{resolve}")
-
-    digest_contents = await Get(
-        DigestContents,
-        PathGlobs(
-            [requirements_full_path],
-            glob_match_error_behavior=GlobMatchErrorBehavior.error,
-            description_of_origin=f"{generator}'s field `{PythonRequirementsSourceField.alias}`",
-        ),
-    )
-    requirements = parse_requirements_file(
-        digest_contents[0].content.decode(), rel_path=requirements_full_path
-    )
-    grouped_requirements = itertools.groupby(
-        requirements, lambda parsed_req: parsed_req.project_name
-    )
-
-    module_mapping = generator[ModuleMappingField].value
-    stubs_mapping = generator[TypeStubsModuleMappingField].value
-
-    def generate_tgt(
-        project_name: str, parsed_reqs: Iterable[PipRequirement]
-    ) -> PythonRequirementTarget:
-        normalized_proj_name = canonicalize_project_name(project_name)
-        tgt_overrides = overrides.pop(normalized_proj_name, {})
-        if Dependencies.alias in tgt_overrides:
-            tgt_overrides[Dependencies.alias] = list(tgt_overrides[Dependencies.alias]) + req_deps
-
-        return PythonRequirementTarget(
-            {
-                **request.template,
-                PythonRequirementsField.alias: list(parsed_reqs),
-                PythonRequirementModulesField.alias: module_mapping.get(normalized_proj_name),
-                PythonRequirementTypeStubModulesField.alias: stubs_mapping.get(
-                    normalized_proj_name
-                ),
-                # This may get overridden by `tgt_overrides`, which will have already added in
-                # the file tgt.
-                Dependencies.alias: req_deps,
-                **tgt_overrides,
-            },
-            request.template_address.create_generated(project_name),
-            union_membership,
-        )
-
-    result = tuple(
-        generate_tgt(project_name, parsed_reqs_)
-        for project_name, parsed_reqs_ in grouped_requirements
-    ) + (file_tgt,)
-
-    if overrides:
-        raise InvalidFieldException(
-            softwrap(
-                f"""
-                Unused key in the `overrides` field for {request.template_address}:
-                {sorted(overrides)}
-                """
-            )
-        )
-
-    return GeneratedTargets(generator, result)
+def parse_requirements_callback(file_contents: bytes, file_path: str) -> Iterator[PipRequirement]:
+    return parse_requirements_file(file_contents.decode(), rel_path=file_path)
 
 
 def rules():
